+++ docs/history.md
@@ -63,6 +63,12 @@
 
 * `oauth2`:
   - `OAuth._requestHandlers['2']` is now async.
+
+* `minifier-css`:
+  - `minifyCss` is now async.
+
+* `webapp`:
+  - `WebAppInternals.getBoilerplate` is now async.
   
 ####  Internal API changes